front: move comparison buttons in their own folder
